refactor:修改物品位置变量定义

+++ src/GameSrv/Actor/ActorConst.cs
@@ -46,14 +46,14 @@ namespace GameSrv.Actor {
     }
 
     public class DragonConst {
-        public const int DRAGON_RING_SHAPE = 198;
-        public const int DRAGON_BRACELET_SHAPE = 199;
-        public const int DRAGON_NECKLACE_SHAPE = 200;
-        public const int DRAGON_DRESS_SHAPE = 10;
-        public const int DRAGON_HELMET_SHAPE = 201;
-        public const int DRAGON_WEAPON_SHAPE = 37;
-        public const int DRAGON_BOOTS_SHAPE = 203;
-        public const int DRAGON_BELT_SHAPE = 204;
+        public const byte DRAGON_RING_SHAPE = 198;
+        public const byte DRAGON_BRACELET_SHAPE = 199;
+        public const byte DRAGON_NECKLACE_SHAPE = 200;
+        public const byte DRAGON_DRESS_SHAPE = 10;
+        public const byte DRAGON_HELMET_SHAPE = 201;
+        public const byte DRAGON_WEAPON_SHAPE = 37;
+        public const byte DRAGON_BOOTS_SHAPE = 203;
+        public const byte DRAGON_BELT_SHAPE = 204;
     }
 
     public class ShapeConst {
@@ -85,58 +85,55 @@ namespace GameSrv.Actor {
     }
 
     public class ItemShapeConst {
-        public const int SHAPE_OF_LUCKYLADLE = 5;
+        public const byte SHAPE_OF_LUCKYLADLE = 5;
         /// <summary>
         /// 隐身属性
         /// </summary>
-        public const int RING_TRANSPARENT_ITEM = 111;
+        public const byte RING_TRANSPARENT_ITEM = 111;
         /// <summary>
         /// 传送属性
         /// </summary>
-        public const int RING_SPACEMOVE_ITEM = 112;
+        public const byte RING_SPACEMOVE_ITEM = 112;
         /// <summary>
         /// 麻痹属性
         /// </summary>
-        public const int RING_MAKESTONE_ITEM = 113;
+        public const byte RING_MAKESTONE_ITEM = 113;
         /// <summary>
         /// 复活属性
         /// </summary>
-        public const int RING_REVIVAL_ITEM = 114;
+        public const byte RING_REVIVAL_ITEM = 114;
         /// <summary>
         /// 火焰属性
         /// </summary>
-        public const int RING_FIREBALL_ITEM = 115;
+        public const byte RING_FIREBALL_ITEM = 115;
         /// <summary>
         /// 治愈属性
         /// </summary>
-        public const int RING_HEALING_ITEM = 116;
+        public const byte RING_HEALING_ITEM = 116;
         /// <summary>
         /// 未知
         /// </summary>
-        public const int RING_ANGERENERGY_ITEM = 117;
+        public const byte RING_ANGERENERGY_ITEM = 117;
         /// <summary>
         /// 护身属性
         /// </summary>
-        public const int RING_MAGICSHIELD_ITEM = 118;
+        public const byte RING_MAGICSHIELD_ITEM = 118;
         /// <summary>
         /// 活力属性
         /// </summary>
-        public const int RING_SUPERSTRENGTH_ITEM = 119;
+        public const byte RING_SUPERSTRENGTH_ITEM = 119;
         /// <summary>
         /// 技巧项链
         /// </summary>
-        public const int NECTLACE_FASTTRAINING_ITEM = 120;
+        public const byte NECTLACE_FASTTRAINING_ITEM = 120;
         /// <summary>
         /// 探测项链
         /// </summary>
-        public const int NECTLACE_SEARCH_ITEM = 121;
-        /// <summary>
-        /// 
-        /// </summary>
-        public const int RING_CHUN_ITEM = 122;
-        public const int NECKLACE_GI_ITEM = 123;
-        public const int ARMRING_HAP_ITEM = 124;
-        public const int HELMET_IL_ITEM = 125;
+        public const byte NECTLACE_SEARCH_ITEM = 121;
+        public const byte RING_CHUN_ITEM = 122;
+        public const byte NECKLACE_GI_ITEM = 123;
+        public const byte ARMRING_HAP_ITEM = 124;
+        public const byte HELMET_IL_ITEM = 125;
 
         /// <summary>
         /// 祈祷套装
@@ -146,83 +143,83 @@ namespace GameSrv.Actor {
         public const byte SpiritItem2 = 128;
         public const byte SpiritItem4 = 129;
 
-        public const int RING_OF_UNKNOWN = 130;
-        public const int BRACELET_OF_UNKNOWN = 131;
-        public const int HELMET_OF_UNKNOWN = 132;
-        public const int RING_OF_MANATOHEALTH = 133;
-        public const int BRACELET_OF_MANATOHEALTH = 134;
+        public const byte RING_OF_UNKNOWN = 130;
+        public const byte BRACELET_OF_UNKNOWN = 131;
+        public const byte HELMET_OF_UNKNOWN = 132;
+        public const byte RING_OF_MANATOHEALTH = 133;
+        public const byte BRACELET_OF_MANATOHEALTH = 134;
         /// <summary>
         /// 魔血项链
         /// </summary>
-        public const int NECKLACE_OF_MANATOHEALTH = 135;
-        public const int RING_OF_SUCKHEALTH = 136;
-        public const int BRACELET_OF_SUCKHEALTH = 137;
+        public const byte NECKLACE_OF_MANATOHEALTH = 135;
+        public const byte RING_OF_SUCKHEALTH = 136;
+        public const byte BRACELET_OF_SUCKHEALTH = 137;
         /// <summary>
         /// 虹膜项链
         /// </summary>
-        public const int NECKLACE_OF_SUCKHEALTH = 138;
-        public const int RING_OF_HPUP = 140;
-        public const int BRACELET_OF_HPUP = 141;
-        public const int RING_OF_MPUP = 142;
-        public const int BRACELET_OF_MPUP = 143;
-        public const int RING_OF_HPMPUP = 144;
-        public const int BRACELET_OF_HPMPUP = 145;
-        public const int NECKLACE_OF_HPPUP = 146;
-        public const int BRACELET_OF_HPPUP = 147;
-        public const int RING_OH_HPPUP = 148;
-        public const int CCHO_WEAPON = 23;
-        public const int CCHO_NECKLACE = 149;
-        public const int CCHO_RING = 150;
-        public const int CCHO_HELMET = 151;
-        public const int CCHO_BRACELET = 152;
-        public const int PSET_NECKLACE_SHAPE = 153;
-        public const int PSET_BRACELET_SHAPE = 154;
-        public const int PSET_RING_SHAPE = 155;
-        public const int HSET_NECKLACE_SHAPE = 156;
-        public const int HSET_BRACELET_SHAPE = 157;
-        public const int HSET_RING_SHAPE = 158;
-        public const int YSET_NECKLACE_SHAPE = 159;
-        public const int YSET_BRACELET_SHAPE = 160;
-        public const int YSET_RING_SHAPE = 161;
-        public const int BONESET_WEAPON_SHAPE = 4;
-        public const int BONESET_HELMET_SHAPE = 162;
-        public const int BONESET_DRESS_SHAPE = 2;
-        public const int BUGSET_NECKLACE_SHAPE = 163;
-        public const int BUGSET_RING_SHAPE = 164;
-        public const int BUGSET_BRACELET_SHAPE = 165;
-        public const int PTSET_BELT_SHAPE = 166;
-        public const int PTSET_BOOTS_SHAPE = 167;
-        public const int PTSET_NECKLACE_SHAPE = 168;
-        public const int PTSET_BRACELET_SHAPE = 169;
-        public const int PTSET_RING_SHAPE = 170;
-        public const int KSSET_BELT_SHAPE = 176;
-        public const int KSSET_BOOTS_SHAPE = 177;
-        public const int KSSET_NECKLACE_SHAPE = 178;
-        public const int KSSET_BRACELET_SHAPE = 179;
-        public const int KSSET_RING_SHAPE = 180;
-        public const int RUBYSET_BELT_SHAPE = 171;
-        public const int RUBYSET_BOOTS_SHAPE = 172;
-        public const int RUBYSET_NECKLACE_SHAPE = 173;
-        public const int RUBYSET_BRACELET_SHAPE = 174;
-        public const int RUBYSET_RING_SHAPE = 175;
-        public const int STRONG_PTSET_BELT_SHAPE = 181;
-        public const int STRONG_PTSET_BOOTS_SHAPE = 182;
-        public const int STRONG_PTSET_NECKLACE_SHAPE = 183;
-        public const int STRONG_PTSET_BRACELET_SHAPE = 184;
-        public const int STRONG_PTSET_RING_SHAPE = 185;
-        public const int STRONG_KSSET_BELT_SHAPE = 191;
-        public const int STRONG_KSSET_BOOTS_SHAPE = 192;
-        public const int STRONG_KSSET_NECKLACE_SHAPE = 193;
-        public const int STRONG_KSSET_BRACELET_SHAPE = 194;
-        public const int STRONG_KSSET_RING_SHAPE = 195;
-        public const int STRONG_RUBYSET_BELT_SHAPE = 186;
-        public const int STRONG_RUBYSET_BOOTS_SHAPE = 187;
-        public const int STRONG_RUBYSET_NECKLACE_SHAPE = 188;
-        public const int STRONG_RUBYSET_BRACELET_SHAPE = 189;
-        public const int STRONG_RUBYSET_RING_SHAPE = 190;
-        public const int DRESS_SHAPE_WING = 9;
-        public const int DRESS_SHAPE_PBKING = 11;
-        public const int DRESS_STDMODE_MAN = 10;
-        public const int DRESS_STDMODE_WOMAN = 11;
+        public const byte NECKLACE_OF_SUCKHEALTH = 138;
+        public const byte RING_OF_HPUP = 140;
+        public const byte BRACELET_OF_HPUP = 141;
+        public const byte RING_OF_MPUP = 142;
+        public const byte BRACELET_OF_MPUP = 143;
+        public const byte RING_OF_HPMPUP = 144;
+        public const byte BRACELET_OF_HPMPUP = 145;
+        public const byte NECKLACE_OF_HPPUP = 146;
+        public const byte BRACELET_OF_HPPUP = 147;
+        public const byte RING_OH_HPPUP = 148;
+        public const byte CCHO_WEAPON = 23;
+        public const byte CCHO_NECKLACE = 149;
+        public const byte CCHO_RING = 150;
+        public const byte CCHO_HELMET = 151;
+        public const byte CCHO_BRACELET = 152;
+        public const byte PSET_NECKLACE_SHAPE = 153;
+        public const byte PSET_BRACELET_SHAPE = 154;
+        public const byte PSET_RING_SHAPE = 155;
+        public const byte HSET_NECKLACE_SHAPE = 156;
+        public const byte HSET_BRACELET_SHAPE = 157;
+        public const byte HSET_RING_SHAPE = 158;
+        public const byte YSET_NECKLACE_SHAPE = 159;
+        public const byte YSET_BRACELET_SHAPE = 160;
+        public const byte YSET_RING_SHAPE = 161;
+        public const byte BONESET_WEAPON_SHAPE = 4;
+        public const byte BONESET_HELMET_SHAPE = 162;
+        public const byte BONESET_DRESS_SHAPE = 2;
+        public const byte BUGSET_NECKLACE_SHAPE = 163;
+        public const byte BUGSET_RING_SHAPE = 164;
+        public const byte BUGSET_BRACELET_SHAPE = 165;
+        public const byte PTSET_BELT_SHAPE = 166;
+        public const byte PTSET_BOOTS_SHAPE = 167;
+        public const byte PTSET_NECKLACE_SHAPE = 168;
+        public const byte PTSET_BRACELET_SHAPE = 169;
+        public const byte PTSET_RING_SHAPE = 170;
+        public const byte KSSET_BELT_SHAPE = 176;
+        public const byte KSSET_BOOTS_SHAPE = 177;
+        public const byte KSSET_NECKLACE_SHAPE = 178;
+        public const byte KSSET_BRACELET_SHAPE = 179;
+        public const byte KSSET_RING_SHAPE = 180;
+        public const byte RUBYSET_BELT_SHAPE = 171;
+        public const byte RUBYSET_BOOTS_SHAPE = 172;
+        public const byte RUBYSET_NECKLACE_SHAPE = 173;
+        public const byte RUBYSET_BRACELET_SHAPE = 174;
+        public const byte RUBYSET_RING_SHAPE = 175;
+        public const byte STRONG_PTSET_BELT_SHAPE = 181;
+        public const byte STRONG_PTSET_BOOTS_SHAPE = 182;
+        public const byte STRONG_PTSET_NECKLACE_SHAPE = 183;
+        public const byte STRONG_PTSET_BRACELET_SHAPE = 184;
+        public const byte STRONG_PTSET_RING_SHAPE = 185;
+        public const byte STRONG_KSSET_BELT_SHAPE = 191;
+        public const byte STRONG_KSSET_BOOTS_SHAPE = 192;
+        public const byte STRONG_KSSET_NECKLACE_SHAPE = 193;
+        public const byte STRONG_KSSET_BRACELET_SHAPE = 194;
+        public const byte STRONG_KSSET_RING_SHAPE = 195;
+        public const byte STRONG_RUBYSET_BELT_SHAPE = 186;
+        public const byte STRONG_RUBYSET_BOOTS_SHAPE = 187;
+        public const byte STRONG_RUBYSET_NECKLACE_SHAPE = 188;
+        public const byte STRONG_RUBYSET_BRACELET_SHAPE = 189;
+        public const byte STRONG_RUBYSET_RING_SHAPE = 190;
+        public const byte DRESS_SHAPE_WING = 9;
+        public const byte DRESS_SHAPE_PBKING = 11;
+        public const byte DRESS_STDMODE_MAN = 10;
+        public const byte DRESS_STDMODE_WOMAN = 11;
     }
 }
